+++ mult.ini
@@ -1,0 +1,10 @@
+[http]
+name="http1"
+port=80
+endpoint = "Endpoint"
+[http]
+name="http2"
+port = 80 
+endpoint = "NotAnEndpoint"
+[ssh]
+port = 1334
